Rename v3 API components to canonical warthog.core module
Fixes #18
