+++ warthog/api.py
@@ -34,12 +34,9 @@ from .transport import get_transport_factory
 from .exceptions import (
     WarthogError,
     WarthogApiError,
-    WarthogAuthCloseError,
     WarthogAuthFailureError,
     WarthogInvalidSessionError,
     WarthogNodeError,
-    WarthogNodeDisableError,
-    WarthogNodeEnableError,
     WarthogNodeStatusError,
     WarthogNoSuchNodeError,
     WarthogPermissionError,
@@ -70,12 +67,9 @@ __all__ = [
     # warthog.exceptions
     'WarthogError',
     'WarthogApiError',
-    'WarthogAuthCloseError',
     'WarthogAuthFailureError',
     'WarthogInvalidSessionError',
     'WarthogNodeError',
-    'WarthogNodeDisableError',
-    'WarthogNodeEnableError',
     'WarthogNodeStatusError',
     'WarthogNoSuchNodeError',
     'WarthogPermissionError',
